% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hullBase-package.R
\docType{package}
\name{hullBase-package}
\alias{hullBase}
\alias{hullBase-package}
\title{hullBase: Rubberband Baseline Correction with Automatic Local Bending}
\description{
Corrects the baseline of spectra, such as Fourier transform infrared (FTIR) spectra, using the rubberband (lower convex hull) method. A rubberband baseline cannot follow concave regions of the background. This package detects hull segments that hide a concave region and corrects each one by bending it locally before taking the hull, extending the manual bending approach described by Beleites (2015). Where to bend and how strongly are calculated from the spectrum's own noise level, so the only user setting is the expected peak width.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThanyaManivannan/hullBase}
  \item Report bugs at \url{https://github.com/ThanyaManivannan/hullBase/issues}
}

}
\author{
\strong{Maintainer}: Thanya Manivannan \email{thanyamanivannan@gmail.com}

Authors:
\itemize{
  \item Thanya Manivannan \email{thanyamanivannan@gmail.com}
}

}
