% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubberband_baseline.R
\name{rubberband_baseline}
\alias{rubberband_baseline}
\title{Create a Rubberband Baseline Object}
\usage{
rubberband_baseline(x, y)
}
\arguments{
\item{x}{Numeric vector of wavenumbers (or any x-axis values).}

\item{y}{Numeric vector of absorbance values, the same length as \code{x}.}
}
\value{
An object of class \code{rubberband_baseline}, storing the sorted x and y values.
}
\description{
Constructs an object representing a spectrum's rubberband baseline model,
to be used with \code{correct()}.
}
\examples{
model <- rubberband_baseline(c(1, 2, 3, 4, 5), c(5, 2, 3, 2, 5))

}
