% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hullBase-package.R
\docType{package}
\name{hullBase-package}
\alias{hullBase}
\alias{hullBase-package}
\title{hullBase: Adaptive, Concave-Aware Rubberband Baseline Correction}
\description{
Provides an automatic, concave-aware extension of the classic rubberband (convex hull) method for baseline correction of spectra such as Fourier-transform infrared (FTIR) data. Standard rubberband baselines cannot represent concave regions of the true background. This package detects concave hull segments automatically and refines them using thresholds derived from each spectrum's own noise level, removing the need for user-supplied tuning parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThanyaManivannan/hullBase}
  \item Report bugs at \url{https://github.com/ThanyaManivannan/hullBase/issues}
}

}
\author{
\strong{Maintainer}: Thanya Manivannan \email{thanyamanivannan@gmail.com}

Authors:
\itemize{
  \item Thanya Manivannan \email{thanyamanivannan@gmail.com}
}

}
\keyword{internal}
