% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.BaselineModel}
\alias{print.summary.BaselineModel}
\title{Print a BaselineModel Summary}
\usage{
\method{print}{summary.BaselineModel}(x, ...)
}
\arguments{
\item{x}{A \code{summary.BaselineModel} object from \code{summary()}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a BaselineModel Summary
}
